refactor: rewrite crawler.js as thin orchestration layer using extracted modules
