+++ backend/src/crawler.js
@@ -1,206 +1,37 @@
 /**
- * crawler.js — Sentri autonomous QA pipeline
+ * crawler.js — Sentri autonomous QA pipeline (thin orchestration layer)
  *
  * 8-layer pipeline:
- *   1. Smart crawl           (pipeline/smartCrawl.js)
+ *   1. Smart crawl           (pipeline/crawlBrowser.js)
  *   2. Element filtering     (pipeline/elementFilter.js)
  *   3. Intent classification (pipeline/intentClassifier.js)
  *   4. Journey generation    (pipeline/journeyGenerator.js)
- *   5. Deduplication         (pipeline/deduplicator.js)
- *   6. Assertion enhancement (pipeline/assertionEnhancer.js)
- *   7. Validate generated tests (syntax + structure checks)
+ *   5. Deduplication         (pipeline/pipelineOrchestrator.js)
+ *   6. Assertion enhancement (pipeline/pipelineOrchestrator.js)
+ *   7. Validate generated tests (pipeline/pipelineOrchestrator.js)
  *   8. Feedback loop         (pipeline/feedbackLoop.js — runs post-execution)
+ *
+ * Sub-concerns extracted to focused modules:
+ *   - pipeline/pageSnapshot.js       — takeSnapshot()
+ *   - pipeline/testValidator.js      — validateTest()
+ *   - pipeline/testPersistence.js    — persistGeneratedTests(), buildPipelineStats()
+ *   - pipeline/crawlBrowser.js       — crawlPages()
+ *   - pipeline/pipelineOrchestrator.js — runPostGenerationPipeline()
  */
 
-import { chromium } from "playwright";
-import { generateTestId } from "./utils/idGenerator.js";
-import { getProviderName, streamText } from "./aiProvider.js";
+import { getProviderName } from "./aiProvider.js";
 import { throwIfAborted, finalizeRunIfNotAborted } from "./utils/abortHelper.js";
-import { SmartCrawlQueue, fingerprintStructure, extractPathPattern } from "./pipeline/smartCrawl.js";
-import { filterElements, hasHighValueElements, filterStats } from "./pipeline/elementFilter.js";
-import { classifyPage, classifyPageWithAI, buildUserJourneys } from "./pipeline/intentClassifier.js";
+import { filterElements, filterStats } from "./pipeline/elementFilter.js";
+import { classifyPageWithAI, buildUserJourneys } from "./pipeline/intentClassifier.js";
 import { generateAllTests, generateUserRequestedTest } from "./pipeline/journeyGenerator.js";
-import { deduplicateTests, deduplicateAcrossRuns } from "./pipeline/deduplicator.js";
-import { enhanceTests } from "./pipeline/assertionEnhancer.js";
-import { emitRunEvent, log, logWarn, logError, logSuccess } from "./utils/runLogger.js";
-
-const MAX_PAGES = parseInt(process.env.CRAWL_MAX_PAGES, 10) || 30;
-const MAX_DEPTH = parseInt(process.env.CRAWL_MAX_DEPTH, 10) || 3;
+import { crawlPages } from "./pipeline/crawlBrowser.js";
+import { runPostGenerationPipeline } from "./pipeline/pipelineOrchestrator.js";
+import { persistGeneratedTests, buildPipelineStats } from "./pipeline/testPersistence.js";
+import { emitRunEvent, log, logWarn, logSuccess } from "./utils/runLogger.js";
 
 function setStep(run, step) {
   run.currentStep = step;
-  // Broadcast a snapshot so the frontend pipeline progress bar updates live
   emitRunEvent(run.id, "snapshot", { run });
-}
-
-// ── Test validation ───────────────────────────────────────────────────────────
-// Rejects malformed or placeholder tests before they enter the DB.
-// Returns an array of issue strings — empty means the test is valid.
-
-function validateTest(test, projectUrl) {
-  const issues = [];
-
-  // Must have a meaningful name
-  if (!test.name || test.name.trim().length < 5) {
-    issues.push("name is missing or too short");
-  }
-
-  // Must have at least one step
-  if (!Array.isArray(test.steps) || test.steps.length === 0) {
-    issues.push("no test steps defined");
-  }
-
-  // Playwright code: if present, must be parseable (contain `async` and braces)
-  if (test.playwrightCode) {
-    if (!test.playwrightCode.includes("async")) {
-      issues.push("playwrightCode missing async function");
-    }
-    if (!test.playwrightCode.includes("{")) {
-      issues.push("playwrightCode missing function body");
-    }
-    // Reject placeholder URLs that the AI sometimes hallucinates
-    if (test.playwrightCode.includes("https://example.com") ||
-        test.playwrightCode.includes("http://example.com")) {
-      issues.push("playwrightCode uses placeholder example.com URL");
-    }
-    // Must reference the actual project URL (or at least page.goto)
-    if (!test.playwrightCode.includes("page.goto")) {
-      issues.push("playwrightCode missing page.goto navigation");
-    }
-  }
-
-  // Reject tests with duplicate/generic names the AI sometimes produces
-  const genericNames = ["test 1", "test 2", "test 3", "untitled", "sample test", "example test"];
-  if (test.name && genericNames.includes(test.name.toLowerCase().trim())) {
-    issues.push("generic placeholder test name");
-  }
-
-  return issues;
-}
-
-const CRAWL_NETWORKIDLE_TIMEOUT = parseInt(process.env.CRAWL_NETWORKIDLE_TIMEOUT, 10) || 5000;
-
-async function takeSnapshot(page) {
-  // Wait for SPA content to settle — domcontentloaded fires too early for SPAs.
-  // Try networkidle first (best for SPAs), fall back to a generous timeout.
-  await page.waitForLoadState("networkidle", { timeout: CRAWL_NETWORKIDLE_TIMEOUT }).catch(() => {});
-
-  return page.evaluate(() => {
-    // Compute the effective ARIA role of an element (explicit or implicit)
-    function getComputedRole(el) {
-      const explicit = el.getAttribute("role");
-      if (explicit) return explicit;
-      const tag = el.tagName.toLowerCase();
-      const type = (el.getAttribute("type") || "").toLowerCase();
-      if (tag === "button") return "button";
-      if (tag === "a" && el.getAttribute("href")) return "link";
-      if (tag === "input") {
-        if (type === "search") return "searchbox";
-        if (type === "checkbox") return "checkbox";
-        if (type === "radio") return "radio";
-        if (type === "submit" || type === "button") return "button";
-        return "textbox";
-      }
-      if (tag === "select") return "combobox";
-      if (tag === "textarea") return "textbox";
-      return "";
-    }
-
-    // ── Capture form structures with field relationships ──────────────────
-    // This gives the AI context about which fields belong to which form,
-    // enabling it to generate tests that fill forms correctly rather than
-    // guessing field order from a flat element list.
-    const formStructures = [];
-    document.querySelectorAll("form").forEach((form, idx) => {
-      const fields = [];
-      form.querySelectorAll("input, select, textarea").forEach(field => {
-        if (field.type === "hidden") return;
-        const label = field.labels?.[0]?.innerText?.trim()
-          || field.getAttribute("aria-label")
-          || field.getAttribute("placeholder")
-          || field.getAttribute("name")
-          || "";
-        fields.push({
-          tag: field.tagName.toLowerCase(),
-          type: field.getAttribute("type") || "",
-          label: label.slice(0, 60),
-          name: field.getAttribute("name") || "",
-          required: field.required || field.getAttribute("aria-required") === "true",
-          testId: field.getAttribute("data-testid") || field.getAttribute("data-cy") || "",
-        });
-      });
-      const submitBtn = form.querySelector('button[type="submit"], input[type="submit"], button:not([type])');
-      formStructures.push({
-        id: form.id || `form-${idx}`,
-        action: form.action || "",
-        method: form.method || "get",
-        fields,
-        submitText: (submitBtn?.innerText || submitBtn?.value || "").trim().slice(0, 40),
-      });
-    });
-
-    // ── Capture semantic page sections ────────────────────────────────────
-    const sections = [];
-    document.querySelectorAll("header, nav, main, aside, footer, [role='banner'], [role='navigation'], [role='main'], [role='complementary'], [role='contentinfo']").forEach(el => {
-      const tag = el.tagName.toLowerCase();
-      const role = el.getAttribute("role") || tag;
-      const headings = Array.from(el.querySelectorAll("h1, h2, h3")).map(h => h.innerText.trim()).slice(0, 3);
-      sections.push({ role, headings });
-    });
-
-    // ── Capture interactive elements with richer metadata ─────────────────
-    const elements = [];
-    document.querySelectorAll(
-      "a, button, input, select, textarea, [role='button'], [role='link'], [role='combobox'], [role='searchbox'], [role='tab'], [role='menuitem'], form"
-    ).forEach((el) => {
-      const text = (el.innerText || el.value || el.placeholder || el.getAttribute("aria-label") || "").trim().slice(0, 80);
-      const computedRole = getComputedRole(el);
-      const ariaLabel = el.getAttribute("aria-label") || "";
-      const placeholder = el.getAttribute("placeholder") || "";
-      // Find the closest label for inputs
-      const labelText = el.labels?.[0]?.innerText?.trim() || "";
-      elements.push({
-        tag: el.tagName.toLowerCase(),
-        text,
-        type: el.getAttribute("type") || "",
-        href: el.getAttribute("href") || "",
-        id: el.id || "",
-        name: el.getAttribute("name") || "",
-        role: computedRole,
-        ariaLabel,
-        placeholder,
-        label: labelText.slice(0, 60),
-        testId: el.getAttribute("data-testid") || el.getAttribute("data-cy") || "",
-        visible: el.offsetParent !== null,
-        disabled: el.disabled || el.getAttribute("aria-disabled") === "true",
-        required: el.required || el.getAttribute("aria-required") === "true",
-        // Which form does this element belong to? Helps AI group interactions.
-        formId: el.closest("form")?.id || "",
-      });
-    });
-
-    // ── Capture heading hierarchy for context ─────────────────────────────
-    const headings = Array.from(document.querySelectorAll("h1, h2, h3"))
-      .map(h => ({ level: parseInt(h.tagName[1]), text: h.innerText.trim().slice(0, 60) }))
-      .slice(0, 10);
-
-    return {
-      title: document.title,
-      url: location.href,
-      elements: elements.filter(e => e.visible).slice(0, 100),
-      h1: Array.from(document.querySelectorAll("h1")).map(h => h.innerText).join(" | "),
-      headings,
-      forms: document.querySelectorAll("form").length,
-      formStructures,
-      sections,
-      hasLoginForm: !!document.querySelector("input[type='password']"),
-      // Additional page signals for the AI
-      hasModals: document.querySelectorAll("[role='dialog'], .modal, [aria-modal='true']").length > 0,
-      hasTabs: document.querySelectorAll("[role='tablist'], [role='tab']").length > 0,
-      hasTable: document.querySelectorAll("table, [role='grid']").length > 0,
-      metaDescription: document.querySelector('meta[name="description"]')?.content?.slice(0, 120) || "",
-    };
-  });
 }
 
 /**
@@ -232,9 +63,7 @@ export async function generateSingleTest(project, run, db, { name, description, 
   setStep(run, 3);
   log(run, `⏭️  Step 3 (Classify) — skipped (user already described the intent)`);
 
-  // ── Step 4: Generate ONE focused test via AI ────────────────────────────
-  // Use a dedicated prompt that generates exactly 1 test matching the user's
-  // name + description, instead of the crawl pipeline's generic 5-8 tests.
+  // ── Step 4: Generate focused test(s) via AI ─────────────────────────────
   throwIfAborted(signal);
   setStep(run, 4);
   log(run, `🤖 Generating test from user description...`);
@@ -246,78 +75,17 @@ export async function generateSingleTest(project, run, db, { name, description, 
   }, { dialsPrompt, testCount, signal });
   log(run, `📝 Raw tests generated: ${rawTests.length}`);
 
-  // ── Step 5: Deduplicate ─────────────────────────────────────────────────
-  throwIfAborted(signal);
-  setStep(run, 5);
-  log(run, `🚫 Deduplicating...`);
-  const existingTests = Object.values(db.tests).filter(t => t.projectId === project.id);
-  const { unique, removed, stats: dedupStats } = deduplicateTests(rawTests);
-  const finalTests = deduplicateAcrossRuns(unique, existingTests);
-  log(run, `   ${removed} duplicates removed | ${unique.length - finalTests.length} already exist | ${finalTests.length} new unique tests`);
-
-  // ── Step 6: Enhance assertions ──────────────────────────────────────────
-  throwIfAborted(signal);
-  setStep(run, 6);
-  log(run, `✨ Enhancing assertions...`);
-  // No real snapshots or classified pages — enhanceTests falls back gracefully
-  const snapshotsByUrl = {};
-  const classifiedPagesByUrl = {};
-  const { tests: enhancedTests, enhancedCount } = enhanceTests(finalTests, snapshotsByUrl, classifiedPagesByUrl);
-  log(run, `   ${enhancedCount} tests had assertions strengthened`);
-
-  // ── Step 7: Validate ────────────────────────────────────────────────────
-  throwIfAborted(signal);
-  setStep(run, 7);
-  log(run, `✅ Validating generated tests...`);
-  const validatedTests = [];
-  let rejected = 0;
-  for (const t of enhancedTests) {
-    const issues = validateTest(t, project.url);
-    if (issues.length === 0) {
-      validatedTests.push(t);
-    } else {
-      rejected++;
-      logWarn(run, `Rejected "${t.name || "unnamed"}": ${issues.join("; ")}`);
-    }
-  }
-  log(run, `   ${validatedTests.length} valid | ${rejected} rejected`);
+  // ── Steps 5-7: Dedup → Enhance → Validate (shared pipeline) ────────────
+  const { validatedTests, enhancedTests, rejected, removed, enhancedCount, dedupStats } =
+    await runPostGenerationPipeline(rawTests, project, db, run, { signal });
 
   // ── Step 8: Store & Done ────────────────────────────────────────────────
-  const createdTestIds = [];
-  for (const t of validatedTests) {
-    const testId = generateTestId(db);
-    db.tests[testId] = {
-      ...t,
-      id: testId,
-      projectId: project.id,
-      name: t.name || name,
-      description: t.description || description || "",
-      sourceUrl: t.sourceUrl || project.url,
-      pageTitle: t.pageTitle || project.name,
-      createdAt: new Date().toISOString(),
-      lastResult: null,
-      lastRunAt: null,
-      qualityScore: t._quality || 0,
-      isJourneyTest: t.isJourneyTest || false,
-      journeyType: t.journeyType || null,
-      assertionEnhanced: t._assertionEnhanced || false,
-      reviewStatus: "draft",
-      reviewedAt: null,
-    };
-    run.tests.push(testId);
-    createdTestIds.push(testId);
-  }
+  const createdTestIds = persistGeneratedTests(validatedTests, project, db, run, {
+    name, description, sourceUrl: project.url, pageTitle: project.name,
+  });
 
   run.testsGenerated = run.tests.length;
-  run.pipelineStats = {
-    pagesFound: 0,
-    rawTestsGenerated: rawTests.length,
-    duplicatesRemoved: removed,
-    assertionsEnhanced: enhancedCount,
-    validationRejected: rejected,
-    journeysDetected: 0,
-    averageQuality: dedupStats.averageQuality,
-  };
+  run.pipelineStats = buildPipelineStats({ rawTests, removed, enhancedCount, rejected, dedupStats });
 
   finalizeRunIfNotAborted(run, () => {
     run.finishedAt = new Date().toISOString();
@@ -333,106 +101,18 @@ export async function generateSingleTest(project, run, db, { name, description, 
 }
 
 export async function crawlAndGenerateTests(project, run, db, { dialsPrompt = "", testCount = "auto", signal } = {}) {
-  const browser = await chromium.launch({
-    headless: process.env.BROWSER_HEADLESS !== "false",
-    executablePath: process.env.PLAYWRIGHT_CHROMIUM_EXECUTABLE_PATH || undefined,
-    args: ["--no-sandbox", "--disable-setuid-sandbox", "--disable-dev-shm-usage"],
-  });
-  const context = await browser.newContext({ userAgent: "Mozilla/5.0 (compatible; Sentri/1.0)" });
-
   const runStart = Date.now();
 
-  const crawlQueue = new SmartCrawlQueue(project.url);
-  crawlQueue.enqueue(project.url, 0);
-
-  const snapshots = [];
-  const snapshotsByUrl = {};
-  const pathPatternsSeen = new Set();
-
+  // ── Step 1: Smart crawl ─────────────────────────────────────────────────
   log(run, `🕷️  Starting smart crawl of ${project.url}`);
   log(run, `🤖 AI provider: ${getProviderName()}`);
   setStep(run, 1);
 
-  if (project.credentials?.usernameSelector) {
-    const loginPage = await context.newPage();
-    try {
-      await loginPage.goto(project.url, { timeout: 15000 });
-      await loginPage.fill(project.credentials.usernameSelector, project.credentials.username);
-      await loginPage.fill(project.credentials.passwordSelector, project.credentials.password);
-      await loginPage.click(project.credentials.submitSelector);
-      await loginPage.waitForLoadState("networkidle", { timeout: 10000 }).catch(() => {});
-      log(run, `🔑 Logged in as ${project.credentials.username}`);
-    } catch (e) {
-      logWarn(run, `Login failed: ${e.message}`);
-    }
-    await loginPage.close();
-  }
-
-  while (crawlQueue.hasMore() && crawlQueue.visitedCount < MAX_PAGES) {
-    if (signal?.aborted) { await browser.close(); throwIfAborted(signal); }
-    const item = crawlQueue.dequeue();
-    if (!item) break;
-    const { url, depth } = item;
-
-    crawlQueue.markVisited(url);
-
-    const pathPattern = extractPathPattern(url);
-    if (pathPatternsSeen.has(pathPattern) && depth > 0) {
-      log(run, `⏭️  Skipping duplicate structure: ${url}`);
-      continue;
-    }
-    pathPatternsSeen.add(pathPattern);
-
-    const page = await context.newPage();
-    try {
-      log(run, `📄 Visiting (depth ${depth}): ${url}`);
-      await page.goto(url, { waitUntil: "domcontentloaded", timeout: 15000 });
-      // takeSnapshot() now calls waitForLoadState('networkidle') internally,
-      // so we no longer need the arbitrary 800ms static wait here.
-
-      const snapshot = await takeSnapshot(page);
-
-      const structureFP = fingerprintStructure(snapshot);
-      if (crawlQueue.isStructureDuplicate(structureFP) && depth > 1) {
-        log(run, `⏭️  Skipping duplicate layout: ${url}`);
-        await page.close();
-        continue;
-      }
-      crawlQueue.markStructureSeen(structureFP);
-
-      snapshots.push(snapshot);
-      snapshotsByUrl[url] = snapshot;
-      run.pagesFound = snapshots.length;
-      // Keep run.pages in sync so the frontend site graph updates live
-      run.pages = snapshots.map(s => ({ url: s.url, title: s.title || s.url, status: "crawled" }));
-
-      if (depth < MAX_DEPTH) {
-        const links = await page.$$eval("a[href]", els => els.map(e => e.href));
-        for (const href of links) {
-          try {
-            const u = new URL(href, url);
-            u.hash = "";
-            u.search = "";
-            const normalized = u.toString();
-            if (new URL(normalized).origin === new URL(project.url).origin) {
-              crawlQueue.enqueue(normalized, depth + 1);
-            }
-          } catch {}
-        }
-      }
-    } catch (err) {
-      logWarn(run, `Failed: ${url} — ${err.message}`);
-    } finally {
-      await page.close();
-    }
-  }
-
-  await browser.close();
-  logSuccess(run, `Smart crawl done. ${snapshots.length} unique pages found.`);
+  const { snapshots, snapshotsByUrl } = await crawlPages(project, run, { signal });
 
   throwIfAborted(signal);
 
-  // Layer 1: Element filtering
+  // ── Step 2: Element filtering ───────────────────────────────────────────
   setStep(run, 2);
   log(run, `🔍 Filtering elements (removing noise)...`);
   const filteredSnapshots = snapshots.map(snap => {
